backend/requirements.txt
# Core
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8

# Testing
pytest>=7.4
pytest-django>=4.7
factory-boy>=3.3
